backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
faiss-cpu==1.7.4
google-generativeai==0.3.2
paddlepaddle==2.6.0
paddleocr==2.7.3
pypdf==3.17.4
pdfminer.six==20231228
readability-lxml==0.8.1
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
tenacity==8.2.3
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
pandas==2.1.4
tiktoken==0.5.2
PyMuPDF==1.23.8
Pillow==10.1.0
